+++ packages/components/dropdown/src/DropdownContext.tsx
@@ -43,7 +43,7 @@ export type DropdownContextCommonProps = PropsWithChildren<{
    */
   defaultOpen?: boolean
   /**
-   * Use `state` prop to assign a specific state to the dropdown, choosing from: `error`, `alert` and `success`. By doing so, the outline styles will be updated, and a state indicator will be displayed accordingly.
+   * Use `state` prop to assign a specific state to the dropdown, choosing from: `error`, `alert` and `success`. By doing so, the outline styles will be updated.
    */
   state?: 'error' | 'alert' | 'success'
   /**