+++ packages/components/dropdown/src/DropdownItems.tsx
@@ -1,6 +1,6 @@
 import { useMergeRefs } from '@spark-ui/use-merge-refs'
 import { cx } from 'class-variance-authority'
-import { CSSProperties, forwardRef, ReactNode, type Ref } from 'react'
+import { forwardRef, ReactNode, type Ref, useLayoutEffect, useRef } from 'react'
 
 import { useDropdownContext } from './DropdownContext'
 
@@ -8,6 +8,17 @@ interface ItemsProps {
   children: ReactNode
   className?: string
 }
+
+/**
+ * BUGFIX
+ *
+ * 1. The !pointer-events-auto class is needed to prevent a bug
+ *    which cannot be reproduced when running Storybook locally,
+ *    in scenarios such as when a Dropdown is nested within a Dialog,
+ *    the "props" object, containing styles computed by Radix
+ *    may erroneously contain "pointerEvents = 'none'", while the Dropdown is open,
+ *    making it impossible to select a value using a pointer device
+ */
 
 export const Items = forwardRef(
   ({ children, className, ...props }: ItemsProps, forwardedRef: Ref<HTMLUListElement>) => {
@@ -19,7 +30,18 @@ export const Items = forwardRef(
       },
     })
 
-    const ref = useMergeRefs(forwardedRef, downshiftRef)
+    const innerRef = useRef<HTMLElement>(null)
+
+    const ref = useMergeRefs(forwardedRef, downshiftRef, innerRef)
+
+    useLayoutEffect(() => {
+      if (!hasPopover) return
+      if (!innerRef.current) return
+
+      if (innerRef.current.parentElement) {
+        innerRef.current.parentElement.style.pointerEvents = isOpen ? '' : 'none'
+      }
+    }, [isOpen, hasPopover])
 
     return (
       <ul
@@ -27,7 +49,9 @@ export const Items = forwardRef(
         className={cx(
           className,
           'flex flex-col',
-          isOpen ? 'block' : 'pointer-events-none invisible absolute opacity-0',
+          isOpen
+            ? '!pointer-events-auto block' /* 1 */
+            : 'pointer-events-none invisible absolute opacity-0',
           hasPopover && 'p-lg'
         )}
         {...props}
@@ -41,7 +65,6 @@ export const Items = forwardRef(
          * A solution would be to make an abstraction of `Dialog.Overlay` instead of using the radix one, but that would mean managing body scroll freeze and scrollbar shifting ourselves.
          *
          */
-        style={{ ...(props as { style: CSSProperties }).style, pointerEvents: undefined }}
         data-spark-component="dropdown-items"
       >
         {children}