+++ packages/components/form-field/src/FormFieldStateMessage.tsx
@@ -1,18 +1,19 @@
+import { Icon } from '@spark-ui/icon'
+import { AlertOutline } from '@spark-ui/icons/dist/icons/AlertOutline'
+import { Check } from '@spark-ui/icons/dist/icons/Check'
+import { WarningOutline } from '@spark-ui/icons/dist/icons/WarningOutline'
+import { cx } from 'class-variance-authority'
 import { forwardRef } from 'react'
 
 import { useFormField } from './FormFieldContext'
 import { FormFieldMessage, FormFieldMessageProps } from './FormFieldMessage'
-import {
-  formFieldStateMessageStyles,
-  FormFieldStateMessageStylesProps,
-} from './FormFieldStateMessage.styles'
 
-export interface FormFieldStateMessageProps
-  extends FormFieldMessageProps,
-    FormFieldStateMessageStylesProps {}
+export interface FormFieldStateMessageProps extends FormFieldMessageProps {
+  state: 'error' | 'alert' | 'success'
+}
 
 export const FormFieldStateMessage = forwardRef<HTMLSpanElement, FormFieldStateMessageProps>(
-  ({ className, state, ...others }, ref) => {
+  ({ className, state, children, ...others }, ref) => {
     const field = useFormField()
 
     if (field.state !== state) {
@@ -24,9 +25,31 @@ export const FormFieldStateMessage = forwardRef<HTMLSpanElement, FormFieldStateM
         ref={ref}
         data-spark-component="form-field-state-message"
         aria-live="polite"
-        className={formFieldStateMessageStyles({ className, state })}
+        className={cx(
+          'flex items-center gap-sm',
+          state === 'error' ? 'text-error' : 'text-on-surface/dim-1',
+          className
+        )}
         {...others}
-      />
+      >
+        {state === 'alert' && (
+          <Icon size="sm">
+            <WarningOutline />
+          </Icon>
+        )}
+        {state === 'error' && (
+          <Icon size="sm" intent="error">
+            <AlertOutline />
+          </Icon>
+        )}
+        {state === 'success' && (
+          <Icon size="sm">
+            <Check />
+          </Icon>
+        )}
+
+        {children}
+      </FormFieldMessage>
     )
   }
 )